{{cookiecutter.repo_name}}/requirements/production.txt
# Production requirements
-r base.txt

{% if cookiecutter.heroku == "y" or cookiecutter.heroku == "Y" %}
dj-database-url==0.2.2
gunicorn==18.0
django-storages==1.1.4
boto==2.21.1
gevent==0.13.8
{% endif %}
